sportscenterapis/requirements.txt
asgiref==3.8.1
Django==5.1.7
django-filter==25.1
djangorestframework==3.15.2
drf-yasg==1.21.10
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyMySQL==1.1.1
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.1
uritemplate==4.1.1

cloudinary~=1.43.0